+++ src/component-classes.ts
@@ -1,7 +1,47 @@
 import { breakpoints } from "./breakpoints.ts";
 import type { CssRuleDefinitionSet } from "./types.ts";
 
+const tableRules: CssRuleDefinitionSet = Array.from(new Array(16)).flatMap((
+    _,
+    n,
+) => [[
+    `table-${n + 1}`,
+    [
+        "display: grid",
+        `grid-template-columns: repeat(${n + 1}, auto)`,
+        "border: 1px solid currentColor",
+        "border-radius: var(--ox-base-length)",
+        "overflow: hidden",
+    ],
+], [
+    `table-${n + 1} > *`,
+    [
+        "border-bottom: 1px solid currentColor",
+        "border-right: 1px solid currentColor",
+        "display: flex",
+        "align-items: center",
+        "padding: var(--ox-base-length)",
+        "min-height: calc(6 * var(--ox-base-length))",
+    ],
+], [
+    `table-${n + 1} *:nth-of-type(${n + 1}n)`,
+    [
+        "border-right: none",
+    ],
+], [
+    `table-${n + 1} *:nth-last-of-type(-n + ${n + 1})`,
+    [
+        "border-bottom: none",
+    ],
+], [
+    `table-${n + 1} *:nth-of-type(${n + 2}n):nth-of-type(${n + 2}n + 2)`,
+    [
+        "background: var(--ox-color-3)",
+    ],
+]]);
+
 const cssRules: CssRuleDefinitionSet = [
+    ...tableRules,
     [
         "button",
         [
@@ -9,15 +49,15 @@ const cssRules: CssRuleDefinitionSet = [
             "display: flex",
             "align-items: center",
             "justify-content: center",
-            "border: 1px solid var(--ox-accent)",
+            "border: 1px solid currentColor",
             "border-radius: calc(var(--ox-round-factor) * var(--ox-base-length))",
             "padding: 0 calc(2 * var(--ox-base-length))",
             "height: calc(6 * var(--ox-base-length))",
             "text-decoration: none",
             "font-size: inherit",
             "line-height: 1.5",
-            "color: var(--ox-accent)",
-            "background: var(--ox-accent-bg)",
+            "color: var(--ox-color-1)",
+            "background: var(--ox-color-2)",
             "font-family: inherit",
             "user-select: none",
             "cursor: pointer",
@@ -30,66 +70,12 @@ const cssRules: CssRuleDefinitionSet = [
     ]],
     ["button.active:not(.toggle)", [
         "color: var(--ox-bg)",
-        "background: var(--ox-accent)",
+        "background: var(--ox-color-1)",
     ]],
     ["button.toggle", [
         "padding-right: calc(3 * var(--ox-base-length))",
     ]],
-    [
-        "table",
-        [
-            "border: 1px solid var(--ox-accent)",
-            "display:flex",
-            "flex-direction: column",
-        ],
-    ],
-    [
-        "table > *",
-        [
-            "min-height: calc(6 * var(--ox-base-length))",
-            "border-bottom: 1px solid var(--ox-accent)",
-            "display: flex",
-            "align-items: stretch",
-        ],
-    ],
-    [
-        "table > *:last-child",
-        ["border-bottom: none"],
-    ],
-    [
-        "table > * > *",
-        [
-            "padding: var(--ox-base-length)",
-            "display:flex",
-            "flex-basis: 1px",
-            "flex-grow: 1",
-            "flex-wrap: wrap",
-            "align-items:center",
-        ],
-    ],
-    [
-        "table > * > *:not(:last-child)",
-        ["border-right: 1px solid var(--ox-accent)"],
-    ],
-    [
-        "table > *:nth-child(2n-1)",
-        ["background: var(--ox-blur)"],
-    ],
-    [
-        "table > *:first-child",
-        [
-            "border: none",
-            "font-weight: var(--ox-f-weight-xl)",
-            "background: var(--ox-accent)",
-            "color: var(--ox-bg)",
-        ],
-    ],
-    [
-        "table > *:first-child > *",
-        [
-            "border-color: var(--ox-bg)",
-        ],
-    ],
+
     ["h1", [
         "font-size: var(--ox-f-size-2xl)",
         "font-weight: var(--ox-f-weight-lg)",
@@ -114,7 +100,7 @@ const cssRules: CssRuleDefinitionSet = [
 
 const cssPseudoRules: CssRuleDefinitionSet = [
     ["button:hover", [
-        "background: var(--ox-blur)",
+        "background: var(--ox-color-3)",
     ]],
     ["button.active:hover:not(.toggle)", [
         "filter: brightness(0.8)",
@@ -127,16 +113,16 @@ const cssPseudoRules: CssRuleDefinitionSet = [
         "position: absolute",
         "top: var(--ox-base-length)",
         "right: var(--ox-base-length)",
-        "border: 1px solid var(--ox-accent)",
+        "border: 1px solid currentColor",
         "width: var(--ox-base-length)",
         "height: var(--ox-base-length)",
         "background: none",
         "border-radius: calc(var(--ox-round-factor) * var(--ox-base-length))",
     ]],
     ["button.toggle.active::after", [
-        "background: var(--ox-accent)",
+        "background: var(--ox-color-1)",
     ]],
-    ["link:hover", ["background: var(--ox-accent)", "color: var(--ox-bg)"]],
+    ["link:hover", ["background: var(--ox-color-1)", "color: var(--ox-bg)"]],
 ];
 
 export function makeComponentClasses(): string {